sna: Destroy the counter damage after a reduction-to-all
If, for instance, we reduce the GPU damage to all we know that there can
be no CPU damage even though it may still have a region with a list of
subtractions. Take advantage of this knowledge and cheaply discard that
damage without having to evaluate it.

This should prevent a paranoid assertion that there is no cpu damage
when discarding the CPU bo for an active pixmap.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>