sna/damage: Fix union of extents with dirty damage but no region
By failing to account for certain paths which would create a damage elt
without fully initialisating the damage region (only the damage extents),
we would later overwrite the damage extents with only the extents for
this operation (rather than the union of this operation with the current
damage). This fixes a regression from 098592ca5d,
(sna: Remove the independent tracking of elts from boxes).

Include the associated damage migration debugging code of the callers.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>