sna: Ensure that the batch mode is always declared before emitting dwords
Initially, the batch->mode was only set upon an actual mode switch,
batch submission would not reset the mode. However, to facilitate fast
ring switching with semaphores, reseting the mode upon batch submission
is desired which means that if we submit the batch in the middle of an
operation we must redeclare its mode before continuing.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>