sna: Add a render ring detiling read path
For SNB, in case you really, really want to use GPU detiling and not
incur the ring switch. Tweaking when to just mmap the target seems to
gain most anyway...

The ulterior motive is that this provides fallback paths for avoiding
the use of TILING_Y with GTT mmaps which is broken on 855gm.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>