sna: Almagamate writes based on the total number of bytes written
Cachelines will only be dirtied for the bytes accessed so a better
metric would based on the total number of pages brought into the TLB
and the total number of cachelines used. Base the decision on whether
to try and amalgamate the upload with others on the number of bytes
copied rather than the overall extents.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>