sna: Do not discard CPU damage for a partial copy
If we are copying a region that does not fill its extents (i.e. is not
singular) then we must be care not to discard the CPU damage that is not
overwritten by the copy.

Fixes regression from 77ee92248582d65a03619d1bb1d93a74468eea00
(sna: Use full usage flags for moving the dst pixmap for a copy).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>