sna: Ensure we trigger a retire for search_linear_cache
Bo used for batch buffers are handled differently and not tracked
through the active cache, so we failed to notice when we might be able
to run retire and recover a suitable buffer for reuse. So simply always
run retire when we might need to create a new linear buffer.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>