sna: Prevent backing pixmaps being created later
We used to allow the backing pixmap to be created later in order to
accommodate ShmPixmaps and ShmPutImage. However, they are now correctly
handled upfront if we choose to accelerate those paths, and so all
choice over whether to attach to a pixmap are made during creation and
are invariant.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>