sna: Encourage promotion of snooped CPU bo to real GPU bo
This fixes the regression in performance of fishietank on gen2. As
the texture atlas is too large to be tiled, one might presume that it
has the same performance characteristics as the snooped linear CPU
buffer. It does not. Therefore if we attempt to reuse a vmap bo, promote
it to a full GPU bo. This hopefully gains the benefit of avoiding the
copy for single shot sources, but still gives us the benefit of avoiding
the clflushes.

On the plus side, it does prove that gen2 handles snoopable memory from
both the blitter and the sampler!

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>