r6xx-r9xx: force 1D tiling for buffer with height < 64
Due to some old kernel issue, height is 8 aligned insided the ddx
For buffer with height btw 57 & 63 this lead ddx to believe it can
allocate a 2D tiled surface while mesa will not align height and
will assume 1D tiled leading to disagreement and rendering issue.
This patch force buffer with height < 64 to be 1D tiled.

Signed-off-by: Jerome Glisse <jglisse@redhat.com>