Include config.h in intel_options.c
Commit 8a9a585341e2 ("Only create a single instance of the intel_options
array") moved the definition of intel_options into a separate .c file.
Several of the defined options are #ifdef'd based on the configuration,
but since config.h is never included, the macros being tested are never
defined.  Therefore, none of the configuration-specific options will
ever be available at runtime, even if they should be.

Add an inclusion of config.h so that such configuration-specific options
work again.

Signed-off-by: Nick Bowler <nbowler@draconx.ca>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>