sna: Actually create Y-tiled source pixmaps
An inconsistency highlighted by 7c51cabaecac revealed that we had a
mismatch between the check in move_to_gpu() and how we created the
pixmap. This mismatch resulted in us creating and uploading tiled
pixmaps for single shot textures, and the increase aperture pressure was
causing a regression in firefox-fishbowl on pnv, for example.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>