sna: Avoid copying unintialised data during source picture upload
If we have never written to a pixmap, then there will be neither a GPU
or shadow pointer and we would attempt to copy a NULL pointer. In this
case as the user is expecting to copy unintialised data we are at
liberty to replace those undefined values with the clear color.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>