sna/gen7: prefer using RENDER copy
Further testing and the balance of doubt swings in favour of using the
3D pipeline for copies.

For small copies the BLT unit is faster,
2.14M/sec vs 1.71M/sec for comppixwin10

And for large copies the RENDER pipeline is faster,
13000/sec vs 8000/sec for comppixwin500

I think the implication is that we are not efficiently utilising the EU
for small primitives - i.e. something that we might be able to improve.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>