xeglgears: Make EGL_KHR_image usage portable
EGL extension functions don't have to be exported which means
xeglgears was failing to link against implementations that
support EGL_KHR_image but were not exporting its related functions.

This has been fixed by using eglGetProcAddress to get a function
pointer instead of using the functions prototype. This is portable.

Signed-off-by: Frank Binns <frank.binns@imgtec.com>