sna: Restore inplace upload for DRI compositors
With a DRI compositor we have to flush the output after every request,
which leads to major inefficiencies. This can be mitigated slightly if
we know we will have to upload shortly, which we track using the pixmap
flush flag.

In particular PutImage is meant to upload inplace to an active DRI
buffer, however this was accidentally dropped in

commit a253c95ec63b2b075e66ae7380fed6a73469eba5
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Sun Jul 15 13:32:35 2012 +0100

    sna: Prefer uploads to be staged in snoopable bo

Performace of putimage500 on ivb i7-3720qm:
  bare:        4610
  gnome-shell: 3000
  patched:     3390

Reported-by: Michael Larabel <Michael@phoronix.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>