sna/gen4+: Implement an opacity shader
Avoid the cumbersome lookup through the alpha gradient texture and
simply multiply the incoming opacity value. The next step will be to
reduce the number of floats required per vertex.

Now that we have removed the primary user of the alpha solid cache, it
may be time to retire that as well.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>