sna: Flush the batch if it contains any DRI pixmaps
This fixes a regression from

commit 02963f489b177d0085006753e91e240545933387
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Sun Aug 19 15:45:35 2012 +0100

    sna: Only submit the batch if flushing a DRI client bo

which made the presumption that we called sna_add_flush_pixmap() for
every DRI pixmap that we used. However, that is only called for the
dirty pixmaps, any native exported pixmap only marks the batch as
requiring a flush. So in those cases we always need to submit the batch
if it contains an exported DRI pixmap.

Reported-by: chr.ohm@gmx.net
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=53967
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>