Fix possible_clones computation for shared encoders between outputs
Libdrm's possible_clones is a mask of encoders. Xorg's possible_clones
is a mask of outputs, so we just can't do the following:

  output->possible_clones = kencoder->possible_clones;

This is a problem on Haswell because, at least with the current
patches floating on the mailing list, there is more than one connector
per encoder.

This patch writes the code to properly translate libdrm's encoder mask
into Xorg's output mask.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>