radeon/radeon_platform_probe: fix Zaphod mode breakage
Using radeon_platform_probe function breaks the Zaphod mode because
it attempts to call xf86AddEntityToScreen multiple times, but nobody
calls xf86SetEntityShared prior to that. Consequently, calls for all
but first device instance fail.

Prior to introduction of platform bus, the logic was that the Probe
function would make the entity sharable, which would cause Xserver
to later make it shared prior to adding it to screen. With the
platform bus loading, add to screen happens in the probe
function so we have to make it shared there.

v2: do not make the entity shared if it was not previously marked
    sharable. Should fix:

    https://bugs.freedesktop.org/show_bug.cgi?id=56680
    https://bugs.freedesktop.org/show_bug.cgi?id=56663

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>