sna/gen6: Tweak programming scanline values
The documentation says that both start/end scanline need to be the line
before the desired value, and so to program the first scanline we need
to set it to the last scanline. The docs also say that the lower 3 bits
are ignored, so tweaked the values programmed accordingly with an extra
check that the window is not reduced to 0.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>