Demonstrate two ways of avoiding callbacks in the client API:   1. conn = mir_wait_for_connection(mir_connect(...));   2. conn = mir_wait_connect(...);
I was trying to avoid the redundancy of having multiple functions per
operation in my previous proposal. However since that was disapproved, this
is the next best thing.

Arguably, we don't need mir_wait_connect() because it's just shorthand for:
  conn = mir_wait_for_connection(mir_connect(server, app, nullptr, nullptr));

And even then, we don't need mir_wait_for_connection(). I just added it to
keep the purists happy and avoid the weakly typed alternative:
  conn = (MirConnection*)mir_wait_for(mir_connect(...));