sna/video: Correct interpretation of 'sync'
The 'sync' parameter past along XvPutImage is not a request for us to
perform the operation synchronised to the vblank, rather a hint that the
client wishes the operation to be synchronous. In Xorg, this is
implemented as an immediate ShmCompletion event so make sure we flush
the operation to hw in this case.

Fixes regression from
commit 195a51353c3af7bd253227da5f759f06cea01f73 [2.21.8]
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Tue Apr 9 19:13:46 2013 +0100

    sna/video: Convert to a pure Xv backend

Reported-by: Matti Hamalainen <ccr@tnsp.org>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=65048
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>