DRI2: add vblank extrapolation function
Implement a helper function that will be called when emulating
the running CRTC. The function should be called only when CRTC
is in DPMS-off state. It will look at the vblank count and the
time that was recorded last time the CRTC was running and
calculate how long one must wait (from present time) until
the target_msc is reached if the CRTC were running.

v2: - CRTC-private now stores frame rate instead of nominal
      vblank period.
    - DIX's timer facility can sometimes wake up the scheduled
      functions more than a millisecond earlier. To avoid
      generating an old MSC, we have to add more margin when
      converting the delay in microseconds to milliseconds.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>