Support loading a different application API backend
By default we use the backend functions from
/system/lib/libubuntu_application_api.so.

Programs can select a different backend with $UBUNTU_PLATFORM_API_BACKEND,
which either needs to be a full path or just the file name (then it will be
looked up in the usual library search path, see dlopen(3)).

This will use Android's shared library functions for libraries in /system/, and
the usual libc ones for any other path.