Fix crash if there is no sensor backend
In environments where there are no hardware sensors/Android libs (e. g. desktop
installations) and the simulated sensor backend is not being used, the hybris
bridge does not have any actual backend to load. Consequently, when trying to
create a sensor with e. g. ua_sensors_proximity_new, DLSYM() returns NULL and
IMPLEMENT_FUNCTION0() will crash trying to call the NULL pointer.

Define a new macro IMPLEMENT_CTOR0 which checks if DLSYM returns NULL and
returns NULL in that case, so that the ua_sensors_*_new() functions behave as
expected.

Rename APITest to SimBackendTest, as this adds a new DefaultBackendTest.