Adds a class describing a co-ordinate transformation
geometry::Transformation describes a combination of scaling, translation and rotation. Rotation is limited to rotation around the center of the given object. In an upcoming change this class should be used to describe the position of a surface in a scene, and thus also used to do co-ordinate transformations for hit testing and input dispatching.
The transformation is storesd in a way that trivial cases can be detected, and that tranformation parameters can be read and changed.