Transforming input coordinates according to what is configured at surfaces
This change removes further values from the InputWindowInfo structure and moves
the responsibilities to input::Surface. The calculation of local motion coordinates 
now takes the transformation matrix applied to the surface when input is received.
That change breaks with how PointerCoords works internally, to resolve that this
change also drops the "raw" co-ordinates supplied with each co-ordinate in each event.
To avoid client ABI changes for this, the items in the structure are just renamed.

The change leaves another TODO open. As soon as the renderer applies a different projection
matrix than an orthographic matrix with the parameters (0, 0, screen_width, screen_height),
touch positions and pixel positions will not match. For that use case to work the input
stack needs means to access those parts/information of the scene.