Do not make QtCompositor call QObject belonging to a different thread
Which makes Qt unhappy.

Also, QtCompositor has no business setting QWindow::visible property. That authority
belongs to application code, not to the QPA.

What QtCompositor/QPA can and should set is the window exposure.

NB: that scheme is not fully functional as the process crashes when exiting due to a
Ctrl+C. But at least it doesn't hang on startup anymore and the implementation is
pointing towards the right direction despite being pretty much work-in-progress.