Workaround GL/GLES conflict between Qt and Mir. Split into 3 packages: GL qpa, GLES qpa & QML plugin
Mir is GLES only. Qt on desktop is desktop-GL only. To workaround this, (ab)use the fact that MESA on the desktop allows you to bind GL apis to a GLES context via eglBindAPI.

To use this, need to compile with QPA separately for GL and GLES targets. Thus need to split the packaging into:
 + qtmir-android - qpa plugin for GLES targets
 + qtmir-desktop - qpa plugin for desktop-GL targets (uses MESA hack)
 + qtdeclarative5-qtmir-plugin - the QML plugin

Also stopped unit tests from being installed