Pull out PointerCoordinate struct shared between two files
structs have file scope. This means that a struct declared within a struct is
still visible outside it. In this case event_deprecated.h is included into
event.h and they both define an identical struct called PointerCoordinate.

Pull the definition out and share it. This does not add new API due to the
scoping rule mentioned previously.