Don't let CMake play with rpath, it cuts itself.
So, the awesome thing about rpath is that it PRECEEDS all the other paths.
If, just for the sake of argument, your cross compiles set an rpath-link to
your partial chroot you'd find that you're surprisingly linking against the
version of libmirclient.so.8 found in there.

Fixes the multiple inclusion of protobuf singletons caused when the system
Mir stuff is linking against one version of libmircommon and we're building
a new version.